+++ src/collective/ploneboard/testing.py
@@ -1,7 +1,10 @@
+from plone.app.robotframework.testing import AUTOLOGIN_LIBRARY_FIXTURE
 from plone.app.testing import PloneSandboxLayer
 from plone.app.testing import applyProfile
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import IntegrationTesting
+from plone.app.testing import FunctionalTesting
+from plone.testing import z2
 
 from zope.configuration import xmlconfig
 
@@ -25,4 +28,13 @@ COLLECTIVE_PLONEBOARD_FIXTURE = CollectiveploneboardLayer()
 COLLECTIVE_PLONEBOARD_INTEGRATION_TESTING = IntegrationTesting(
     bases=(COLLECTIVE_PLONEBOARD_FIXTURE,),
     name="CollectiveploneboardLayer:Integration"
+)
+
+COLLECTIVE_PLONEBOARD_ROBOT_TESTING = FunctionalTesting(
+    bases=(
+        COLLECTIVE_PLONEBOARD_FIXTURE,
+        AUTOLOGIN_LIBRARY_FIXTURE,
+        z2.ZSERVER_FIXTURE
+    ),
+    name="CollectivePloneboardLayer:Robot"
 )