+++ CHANGES.rst
@@ -4,8 +4,14 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
-- Package created using templer
-  [timo]
+- Rating comments added
+  [aman]
+
+- Captcha support added
+  [aman]
 
 - Doctests added for workflows
   [aman]
+
+- Package created using templer
+  [timo]